refactor: split session/mod.rs into focused sub-modules
Break the 1,575-line session orchestrator into 4 new files, each
with a clear single concern:

- session/wiring.rs (278 lines): all set/clear/attach methods
- session/frame_processing.rs (303 lines): per-frame render + encode
- session/detection_dispatch.rs (336 lines): AI detection per platform
- session/run_loop.rs (332 lines): batch frame loops (run, run_immediate)

session/mod.rs retains struct definition, constructors, accessors,
and finish (376 lines).

Uses the existing pattern from zero_copy_linux.rs: impl StitchSession
blocks in sub-files accessing pub(crate) fields.

14 tests pass, clippy clean.
